+++ deps/v8/src/base/platform/platform-openbsd.cc
@@ -27,13 +27,16 @@
 #undef MAP_TYPE
 
 #include "src/base/macros.h"
+#include "src/base/platform/platform-posix-time.h"
 #include "src/base/platform/platform-posix.h"
 #include "src/base/platform/platform.h"
 
 namespace v8 {
 namespace base {
 
-TimezoneCache* OS::CreateTimezoneCache() { return new PosixTimezoneCache(); }
+TimezoneCache* OS::CreateTimezoneCache() {
+  return new PosixDefaultTimezoneCache();
+}
 
 void* OS::Allocate(const size_t requested, size_t* allocated,
                    OS::MemoryPermission access) {
