+++ deps/v8/src/base/platform/platform-posix.cc
@@ -388,26 +388,6 @@ double OS::TimeCurrentMillis() {
   return Time::Now().ToJsTime();
 }
 
-#if !V8_OS_AIX && !V8_OS_SOLARIS && !V8_OS_CYGWIN
-const char* PosixTimezoneCache::LocalTimezone(double time) {
-  if (std::isnan(time)) return "";
-  time_t tv = static_cast<time_t>(std::floor(time / msPerSecond));
-  struct tm tm;
-  struct tm* t = localtime_r(&tv, &tm);
-  if (!t || !t->tm_zone) return "";
-  return t->tm_zone;
-}
-
-double PosixTimezoneCache::LocalTimeOffset() {
-  time_t tv = time(NULL);
-  struct tm tm;
-  struct tm* t = localtime_r(&tv, &tm);
-  // tm_gmtoff includes any daylight savings offset, so subtract it.
-  return static_cast<double>(t->tm_gmtoff * msPerSecond -
-                             (t->tm_isdst > 0 ? 3600 * msPerSecond : 0));
-}
-#endif
-
 double PosixTimezoneCache::DaylightSavingsOffset(double time) {
   if (std::isnan(time)) return std::numeric_limits<double>::quiet_NaN();
   time_t tv = static_cast<time_t>(std::floor(time/msPerSecond));
