+++ deps/v8/src/base/platform/platform-posix.h
@@ -13,9 +13,7 @@ namespace base {
 
 class PosixTimezoneCache : public TimezoneCache {
  public:
-  const char* LocalTimezone(double time_ms) override;
   double DaylightSavingsOffset(double time_ms) override;
-  double LocalTimeOffset() override;
   void Clear() override {}
   ~PosixTimezoneCache() override {}
