+++ deps/v8/src/base/platform/platform-qnx.cc
@@ -31,6 +31,7 @@
 #undef MAP_TYPE
 
 #include "src/base/macros.h"
+#include "src/base/platform/platform-posix-time.h"
 #include "src/base/platform/platform-posix.h"
 #include "src/base/platform/platform.h"
 
@@ -84,7 +85,9 @@ bool OS::ArmUsingHardFloat() {
 
 #endif  // __arm__
 
-TimezoneCache* OS::CreateTimezoneCache() { return new PosixTimezoneCache(); }
+TimezoneCache* OS::CreateTimezoneCache() {
+  return new PosixDefaultTimezoneCache();
+}
 
 void* OS::Allocate(const size_t requested, size_t* allocated,
                    OS::MemoryPermission access) {
