+++ deps/v8/src/v8.gyp
@@ -2063,6 +2063,8 @@
               'base/platform/platform-linux.cc',
               'base/platform/platform-posix.h',
               'base/platform/platform-posix.cc',
+              'base/platform/platform-posix-time.h',
+              'base/platform/platform-posix-time.cc',
             ],
           }
         ],
@@ -2071,6 +2073,8 @@
               'base/debug/stack_trace_android.cc',
               'base/platform/platform-posix.h',
               'base/platform/platform-posix.cc',
+              'base/platform/platform-posix-time.h',
+              'base/platform/platform-posix-time.cc',
             ],
             'link_settings': {
               'target_conditions': [
@@ -2127,6 +2131,8 @@
               'base/debug/stack_trace_posix.cc',
               'base/platform/platform-posix.h',
               'base/platform/platform-posix.cc',
+              'base/platform/platform-posix-time.h',
+              'base/platform/platform-posix-time.cc',
               'base/qnx-math.h'
             ],
             'target_conditions': [
@@ -2158,6 +2164,8 @@
               'base/platform/platform-freebsd.cc',
               'base/platform/platform-posix.h',
               'base/platform/platform-posix.cc',
+              'base/platform/platform-posix-time.h',
+              'base/platform/platform-posix-time.cc',
             ],
           }
         ],
@@ -2170,6 +2178,8 @@
               'base/platform/platform-openbsd.cc',
               'base/platform/platform-posix.h',
               'base/platform/platform-posix.cc'
+              'base/platform/platform-posix-time.h',
+              'base/platform/platform-posix-time.cc',
             ],
           }
         ],
@@ -2183,6 +2193,8 @@
               'base/platform/platform-openbsd.cc',
               'base/platform/platform-posix.h',
               'base/platform/platform-posix.cc',
+              'base/platform/platform-posix-time.h',
+              'base/platform/platform-posix-time.cc',
             ],
           }
         ],
@@ -2213,6 +2225,8 @@
             'base/platform/platform-macos.cc',
             'base/platform/platform-posix.h',
             'base/platform/platform-posix.cc',
+            'base/platform/platform-posix-time.h',
+            'base/platform/platform-posix-time.cc',
           ]},
         ],
         ['OS=="win"', {
